moved AuthInterceptor out of keycloak folder
